mv commands package into main > v2ctl is removed, commands is used only by main
