+++ main/commands/all/api/api.go
@@ -1,7 +1,7 @@
 package api
 
 import (
-	"v2ray.com/core/commands/base"
+	"v2ray.com/core/main/commands/base"
 )
 
 // CmdAPI calls an API in an V2Ray process
