+++ main/commands/all/api/inbounds_add.go
@@ -4,18 +4,18 @@ import (
 	"fmt"
 
 	handlerService "v2ray.com/core/app/proxyman/command"
-	"v2ray.com/core/commands/base"
+	"v2ray.com/core/main/commands/base"
 	"v2ray.com/core/common/cmdarg"
 	"v2ray.com/core/infra/conf"
 	"v2ray.com/core/infra/conf/serial"
 )
 
-var cmdAddOutbounds = &base.Command{
+var cmdAddInbounds = &base.Command{
 	CustomFlags: true,
-	UsageLine:   "{{.Exec}} api ado [--server=127.0.0.1:8080] <c1.json> [c2.json]...",
-	Short:       "Add outbounds",
+	UsageLine:   "{{.Exec}} api adi [--server=127.0.0.1:8080] <c1.json> [c2.json]...",
+	Short:       "Add inbounds",
 	Long: `
-Add outbounds to V2Ray.
+Add inbounds to V2Ray.
 
 Arguments:
 
@@ -29,19 +29,19 @@ Example:
 
     {{.Exec}} {{.LongName}} --server=127.0.0.1:8080 c1.json c2.json
 `,
-	Run: executeAddOutbounds,
+	Run: executeAddInbounds,
 }
 
-func executeAddOutbounds(cmd *base.Command, args []string) {
+func executeAddInbounds(cmd *base.Command, args []string) {
 	setSharedFlags(cmd)
 	cmd.Flag.Parse(args)
 	unnamedArgs := cmd.Flag.Args()
 	if len(unnamedArgs) == 0 {
-		fmt.Println("Reading from STDIN")
+		fmt.Println("reading from stdin:")
 		unnamedArgs = []string{"stdin:"}
 	}
 
-	outs := make([]conf.OutboundDetourConfig, 0)
+	ins := make([]conf.InboundDetourConfig, 0)
 	for _, arg := range unnamedArgs {
 		r, err := cmdarg.LoadArg(arg)
 		if err != nil {
@@ -51,28 +51,28 @@ func executeAddOutbounds(cmd *base.Command, args []string) {
 		if err != nil {
 			base.Fatalf("failed to decode %s: %s", arg, err)
 		}
-		outs = append(outs, conf.OutboundConfigs...)
+		ins = append(ins, conf.InboundConfigs...)
 	}
-	if len(outs) == 0 {
-		base.Fatalf("no valid outbound found")
+	if len(ins) == 0 {
+		base.Fatalf("no valid inbound found")
 	}
 
 	conn, ctx, close := dialAPIServer()
 	defer close()
 
 	client := handlerService.NewHandlerServiceClient(conn)
-	for _, out := range outs {
-		fmt.Println("adding:", out.Tag)
-		o, err := out.Build()
+	for _, in := range ins {
+		fmt.Println("adding:", in.Tag)
+		i, err := in.Build()
 		if err != nil {
 			base.Fatalf("failed to build conf: %s", err)
 		}
-		r := &handlerService.AddOutboundRequest{
-			Outbound: o,
+		r := &handlerService.AddInboundRequest{
+			Inbound: i,
 		}
-		resp, err := client.AddOutbound(ctx, r)
+		resp, err := client.AddInbound(ctx, r)
 		if err != nil {
-			base.Fatalf("failed to add outbound: %s", err)
+			base.Fatalf("failed to add inbound: %s", err)
 		}
 		showResponese(resp)
 	}
