+++ main/commands/all/convert_confs.go
@@ -6,7 +6,7 @@ import (
 	"path/filepath"
 	"strings"
 
-	"v2ray.com/core/commands/base"
+	"v2ray.com/core/main/commands/base"
 	"v2ray.com/core/common/cmdarg"
 	"v2ray.com/core/infra/conf/json"
 	"v2ray.com/core/infra/conf/merge"
