+++ main/commands/all/merge_doc.go
@@ -1,7 +1,7 @@
 package all
 
 import (
-	"v2ray.com/core/commands/base"
+	"v2ray.com/core/main/commands/base"
 )
 
 var docMerge = &base.Command{
