+++ main/commands/all/uuid.go
@@ -3,7 +3,7 @@ package all
 import (
 	"fmt"
 
-	"v2ray.com/core/commands/base"
+	"v2ray.com/core/main/commands/base"
 	"v2ray.com/core/common/uuid"
 )
