+++ main/commands/test.go
@@ -5,7 +5,7 @@ import (
 	"log"
 
 	"v2ray.com/core"
-	"v2ray.com/core/commands/base"
+	"v2ray.com/core/main/commands/base"
 )
 
 // CmdTest tests config files
